Refactoring Dashboard components
